requirements.txt
pyserial==3.5
coloredlogs==15.0.1
pywin32